requirements.txt
fastapi
fastapi_utils
uvicorn
python-multipart
pydantic
pydantic[email]
Crypto
python-jose
passlib
sqlalchemy<2.0
mysqlclient
beautifulsoup4
simplejson
requests
alembic
sqlalchemy-utils>=0.36.8
datetime
apscheduler
traceback2